+++ README/CHANGELOG.md
@@ -121,3 +121,13 @@ Initial Release
 
 The next update will focus more on features / compatibility to other systems.
 I hope that most issues are now fixed.
+
+# 2.5
+
+- Added search filters
+- Move settings to a new page of stacked Widget
+- Added keyboard shortcut and some buttons for it
+- API Updated to v3.2
+- Added function for downloading the thumbnail
+- Fixed an error with the metadata function
+- Removed file logging